+++ core-bundle/CHANGELOG.md
@@ -1,3 +1,5 @@
 # Contao core bundle change log
 
 ### DEV
+
+ * Store the request token as cookie instead of in the session (see #1065).